refactor(probe): drop probe Destroy logic
Registering and destroying Probes and QueryExecutors did not add any
real benefits while increasing complexity. This also caused issues when
updating environments as probes already existed and would have to be
carefully updated to not register a probe twice or destroying probes
that are still used.

To keep the logic simple formally registering and destroying probes has
been removed.

Signed-off-by: Simon Berz <simon@berz.me>